+++ src/app/(app)/(main)/why-urlhub/page.tsx
@@ -8,9 +8,9 @@ const WhyUrlHub = () => {
             <section className="text-gray-600 body-font">
                 <div className="flex flex-col-reverse md:flex-row items-center gap-8 md:gap-12 lg:gap-16">
                     <div className="w-full md:w-1/2 flex flex-col items-center md:items-start text-center md:text-left">
-                        <h1 className="text-3xl sm:text-4xl lg:text-5xl font-bold text-gray-900 mb-4">Shorten and Track Your Links Easily</h1>
+                        <h1 className="text-3xl sm:text-4xl lg:text-5xl font-bold text-gray-900 mb-4">Shorten and Track Your Urls Easily</h1>
                         <p className="mb-6 text-base lg:text-lg">
-                            UrlHub allows you to transform long, complex URLs into short, user-friendly links in a matter of seconds. Additionally, you can track every click on your shortened links, providing you with detailed insights into link performance. Whether you&apos;re sharing links on social media, via email, or in presentations, shorter URLs help improve clarity and ensure smoother sharing experiences. With just a few clicks, you&apos;ll have a memorable link that&apos;s ready for distribution across all your channels.
+                            UrlHub allows you to transform long, complex URLs into short, user-friendly urls in a matter of seconds. Additionally, you can track every click on your shortened urls, providing you with detailed insights into url performance. Whether you&apos;re sharing urls on social media, via email, or in presentations, shorter URLs help improve clarity and ensure smoother sharing experiences. With just a few clicks, you&apos;ll have a memorable url that&apos;s ready for distribution across all your channels.
                         </p>
                         <Link href="/register">
                             <button className="w-full md:w-auto text-white bg-black hover:bg-blue-700 focus:ring-blue-300 font-medium rounded-full text-base py-3 px-6 text-center transition duration-300">
@@ -19,7 +19,7 @@ const WhyUrlHub = () => {
                         </Link>
                     </div>
                     <div className="w-full md:w-1/2">
-                        <img className="w-full h-auto object-cover object-center rounded" alt="Shorten and track your links" src='/why1.svg' />
+                        <img className="w-full h-auto object-cover object-center rounded" alt="Shorten and track your urls" src='/why1.svg' />
                     </div>
                     
                 </div>
@@ -33,7 +33,7 @@ const WhyUrlHub = () => {
                     <div className="w-full md:w-1/2 flex flex-col items-center md:items-start text-center md:text-left">
                         <h1 className="text-3xl sm:text-4xl font-bold text-gray-900 mb-4">URL Expiry</h1>
                         <p className="mb-6 text-base lg:text-lg">
-                            UrlHub allows you to set expiry dates for your shortened URLs. This feature ensures that your links are only accessible for a specified period, providing an additional layer of control and security. Whether you&apos;re running time-sensitive promotions or sharing temporary content, URL expiry helps you manage the lifespan of your links effectively.
+                            UrlHub allows you to set expiry dates for your shortened URLs. This feature ensures that your urls are only accessible for a specified period, providing an additional layer of control and security. Whether you&apos;re running time-sensitive promotions or sharing temporary content, URL expiry helps you manage the lifespan of your urls effectively.
                         </p>
                         <Link href="/features">
                             <button className="w-full md:w-auto text-white bg-black hover:bg-blue-700 focus:ring-blue-300 font-medium rounded-full text-base py-3 px-6 text-center transition duration-300">
@@ -49,7 +49,7 @@ const WhyUrlHub = () => {
                     <div className="w-full md:w-1/2 flex flex-col items-center md:items-start text-center md:text-left">
                         <h1 className="text-3xl sm:text-4xl font-bold text-gray-900 mb-4">Password Protected URLs</h1>
                         <p className="mb-6 text-base lg:text-lg">
-                            With UrlHub, you can add password protection to your shortened URLs. This feature ensures that only users with the correct password can access your links, providing an extra layer of security for sensitive information. Whether you&apos;re sharing confidential documents or private content, password-protected URLs help you maintain control over who can view your links.
+                            With UrlHub, you can add password protection to your shortened URLs. This feature ensures that only users with the correct password can access your urls, providing an extra layer of security for sensitive information. Whether you&apos;re sharing confidential documents or private content, password-protected URLs help you maintain control over who can view your urls.
                         </p>
                         <Link href="/dashboard">
                             <button className="w-full md:w-auto text-white bg-black hover:bg-blue-700 focus:ring-blue-300 font-medium rounded-full text-base py-3 px-6 text-center transition duration-300">