+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - Added URL expiry feature to automatically disable links after a certain period.
 - Introduced password-protected URLs to enhance security by requiring a password to access the content.
 - Add new layout components for Pricing, Products, and Why UrlHub pages
+- Updated UI indicators for password protected urls
 
 ### Fixed
 - Make the fun facts random